OneDrive/Desktop/collage management/backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
twilio==8.11.1
sendgrid==6.11.0
python-dotenv==1.0.0
Pillow==10.2.0
